+++ Actionsrecognition/train.py
@@ -1,17 +1,12 @@
 import os
-import time
-import torch
 import pickle
-import numpy as np
-import torch.nn.functional as F
-from shutil import copyfile
 from tqdm import tqdm
 from torch.utils import data
 from torch.optim.adadelta import Adadelta
 from sklearn.model_selection import train_test_split
 
 from Actionsrecognition.Models import *
-from Visualizer import plot_graphs, plot_confusion_metrix
+from Modules.Visualize.Visualizer import plot_graphs, plot_confusion_metrix
 
 
 save_folder = 'saved/TSSTG(pts+mot)-01(cf+hm-hm)'
