+++ Modules/Detect/DetectorLoader.py
